renamed dbtables -> entitites
