+++ src/com/elle/ProjectManager/logic/Issue.java
@@ -1,8 +1,4 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
+
 package com.elle.ProjectManager.logic;
 
 import com.elle.ProjectManager.dao.IssueWindowDAO;
